Rename and combine protocols
