+++ Challo/Challo/Modules/LoginRegister/Login/GuideLogin/GuideLoginAPI.swift
@@ -5,7 +5,7 @@
 //  Created by Tan Le Yang on 19/3/21.
 //
 
-class GuideLoginLogic: LoginLogic {
+class GuideLoginAPI: LoginAPI {
 
     let networkManager = AlamofireManager.alamofireManager
 }
