+++ Challo/Challo/Modules/LoginRegister/Login/TouristLogin/TouristLoginAPI.swift
@@ -5,7 +5,7 @@
 //  Created by Tan Le Yang on 19/3/21.
 //
 
-class TouristLoginLogic: LoginLogic {
+class TouristLoginAPI: LoginAPI {
     
     let networkManager = AlamofireManager.alamofireManager
 }
