+++ Challo/Challo/Modules/LoginRegister/Login/TouristLogin/TouristLoginInteractor.swift
@@ -9,10 +9,10 @@ class TouristLoginInteractor: LoginInteractor, InteractorProtocol {
 
     let networkManager = AlamofireManager.alamofireManager
     weak var presenter: LoginPresenter!
-    private let loginLogic = TouristLoginLogic()
+    private let api = TouristLoginAPI()
 
     func defaultLogin(email: String, password: String) {
-        loginLogic.login(email: email, password: password) { [weak self] response in
+        api.login(email: email, password: password) { [weak self] response in
             if response.error != nil {
                 self?.presenter.showLoginFailureAlert()
                 return
