+++ Challo/Challo/Modules/LoginRegister/Register/GuideRegister/GuideRegistrationAPI.swift
@@ -5,7 +5,7 @@
 //  Created by Tan Le Yang on 19/3/21.
 //
 
-class GuideRegistrationLogic: RegistrationLogic {
+class GuideRegistrationAPI: RegisterAPI {
 
     let userTypeUrl = "/tourist"
     let networkManager = AlamofireManager.alamofireManager
