+++ Challo/Challo/Modules/LoginRegister/Register/TouristRegister/TouristRegisterModule.swift
@@ -18,10 +18,10 @@ final class TouristRegisterModule: ViperModuleProtocol {
         presenter.router = router
         router.presenter = presenter
 
-        let loginLogic = TouristLoginLogic()
-        let registerLogic = TouristRegistrationLogic()
+        let loginAPI = TouristLoginAPI()
+        let registerAPI = TouristRegistrationAPI()
         let socialPresenter = SocialLoginPresenter
-            .createSocialLoginPresenter(loginLogic: loginLogic, registrationLogic: registerLogic)
+            .createSocialLoginPresenter(loginAPI: loginAPI, registerAPI: registerAPI)
 
         let registerView = AnyView(TouristRegisterPage(registerPresenter: presenter,
                                                        socialLoginPresenter: socialPresenter))
